# Conflicts:
#	typescript/common-resolvers/src/dataset.ts

+++ typescript/db/src/repository/index.ts
@@ -1,62 +1,116 @@
-import { PrismaClient, Prisma } from "@prisma/client";
-
+import { Prisma } from "@prisma/client";
 import { DbLabel, Repository } from "@labelflow/common-resolvers";
 import { Image } from "@labelflow/graphql-types";
 import slugify from "slugify";
+import { prisma } from "./prisma-client";
 import {
   getUploadTargetHttp,
   getFromStorage,
   putInStorage,
   deleteFromStorage,
 } from "./upload-supabase";
-import { countLabels, listLabels } from "./label";
+import {
+  addWorkspace,
+  getWorkspace,
+  listWorkspace,
+  updateWorkspace,
+} from "./workspace";
+import { listLabels } from "./label";
 import { castObjectNullsToUndefined } from "./utils";
-
-export const prisma = new PrismaClient();
+import {
+  checkUserAccessToDataset,
+  checkUserAccessToImage,
+  checkUserAccessToLabel,
+  checkUserAccessToLabelClass,
+  checkUserAccessToWorkspace,
+} from "./access-control";
 
 export const repository: Repository = {
   image: {
-    add: async (image) => {
+    add: async (image, user) => {
+      await checkUserAccessToDataset({ where: { id: image.datasetId }, user });
       const createdImage = await prisma.image.create({
         data: castObjectNullsToUndefined(image),
       });
       return createdImage.id;
     },
-    count: (where) =>
-      prisma.image.count({
-        where: castObjectNullsToUndefined(where),
-      }),
-    get: async (where) =>
-      (await prisma.image.findUnique({
+    count: async (whereWithUser) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.image.count({
+        where: castObjectNullsToUndefined({
+          ...where,
+          dataset: {
+            workspace: { memberships: { some: { userId: user?.id } } },
+          },
+        }),
+      });
+    },
+    get: async (where, user) => {
+      await checkUserAccessToImage({ where, user });
+      return (await prisma.image.findUnique({
         where,
-      })) as unknown as Image,
+      })) as unknown as Image;
+    },
 
-    list: (where, skip = undefined, first = undefined) =>
-      prisma.image.findMany(
+    list: async (whereWithUser, skip = undefined, first = undefined) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.image.findMany(
         castObjectNullsToUndefined({
-          where: castObjectNullsToUndefined(where),
+          where: castObjectNullsToUndefined({
+            ...where,
+            dataset: {
+              workspace: { memberships: { some: { userId: user?.id } } },
+            },
+          }),
           orderBy: { createdAt: Prisma.SortOrder.asc },
           skip,
           take: first,
         })
-      ),
+      );
+    },
   },
   label: {
-    add: async (label) => {
+    add: async (label, user) => {
+      await checkUserAccessToImage({ where: { id: label.imageId }, user });
       const createdLabel = await prisma.label.create({ data: label });
       return createdLabel.id;
     },
-    count: countLabels,
-    delete: async ({ id }) => {
+    count: async (whereWithUser) => {
+      if (!whereWithUser) {
+        return 0;
+      }
+      const { user, ...where } = whereWithUser;
+      if ("datasetId" in where) {
+        checkUserAccessToDataset({ where: { id: where.datasetId }, user });
+        return await prisma.label.count({
+          where: { image: { datasetId: where.datasetId ?? undefined } },
+        });
+      }
+      return await prisma.label.count({
+        where: castObjectNullsToUndefined({
+          ...where,
+          image: {
+            dataset: {
+              workspace: { memberships: { some: { userId: user?.id } } },
+            },
+          },
+        }),
+      });
+    },
+    delete: async ({ id }, user) => {
+      await checkUserAccessToLabel({ where: { id }, user });
       await prisma.label.delete({ where: { id } });
     },
     /* Needs to be casted as Prisma doesn't let us specify
      * the type for geometry */
-    get: (where) =>
-      prisma.label.findUnique({
+    get: async (where, user) => {
+      await checkUserAccessToLabel({ where, user });
+      return (await prisma.label.findUnique({
         where,
-      }) as unknown as Promise<DbLabel>,
-    update: async (where, label) => {
+      })) as unknown as DbLabel;
+    },
+    update: async (where, label, user) => {
+      await checkUserAccessToLabel({ where, user });
       try {
         if (label) {
           await prisma.label.update({
@@ -72,34 +126,56 @@ export const repository: Repository = {
     list: listLabels,
   },
   labelClass: {
-    add: async (labelClass) => {
+    add: async (labelClass, user) => {
+      await checkUserAccessToDataset({
+        where: { id: labelClass.datasetId },
+        user,
+      });
       const createdLabelClass = await prisma.labelClass.create({
         data: castObjectNullsToUndefined(labelClass),
       });
       return createdLabelClass.id;
     },
-    count: async (where) =>
-      await prisma.labelClass.count({
-        where: castObjectNullsToUndefined(where),
-      }),
-    delete: async ({ id }) => {
+    count: async (whereWithUser) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.labelClass.count({
+        where: castObjectNullsToUndefined({
+          ...where,
+          dataset: {
+            workspace: { memberships: { some: { userId: user?.id } } },
+          },
+        }),
+      });
+    },
+    delete: async ({ id }, user) => {
+      await checkUserAccessToLabelClass({ where: { id }, user });
       await prisma.labelClass.delete({ where: { id } });
     },
-    get: (where) =>
-      prisma.labelClass.findUnique({
+    get: async (where, user) => {
+      await checkUserAccessToLabelClass({ where, user });
+      return await prisma.labelClass.findUnique({
         where,
-      }),
-    list: (where, skip = undefined, first = undefined) =>
-      prisma.labelClass.findMany(
+      });
+    },
+    list: async (whereWithUser, skip = undefined, first = undefined) => {
+      const { user, ...where } = whereWithUser ?? { user: undefined };
+      return await prisma.labelClass.findMany(
         castObjectNullsToUndefined({
-          where: castObjectNullsToUndefined(where),
+          where: castObjectNullsToUndefined({
+            ...where,
+            dataset: {
+              workspace: { memberships: { some: { userId: user?.id } } },
+            },
+          }),
           orderBy: { index: Prisma.SortOrder.asc },
           skip,
           take: first,
         })
-      ),
-    update: async ({ id }, labelClass) => {
+      );
+    },
+    update: async ({ id }, labelClass, user) => {
       try {
+        await checkUserAccessToLabelClass({ where: { id }, user });
         await prisma.labelClass.update({
           where: { id },
           data: castObjectNullsToUndefined(labelClass),
@@ -111,7 +187,11 @@ export const repository: Repository = {
     },
   },
   dataset: {
-    add: async ({ workspaceSlug, ...dataset }) => {
+    add: async ({ workspaceSlug, ...dataset }, user) => {
+      await checkUserAccessToWorkspace({
+        where: { slug: workspaceSlug },
+        user,
+      });
       const slug = slugify(dataset.name, { lower: true });
       const createdDataset = await prisma.dataset.create({
         data: castObjectNullsToUndefined({
@@ -124,7 +204,7 @@ export const repository: Repository = {
       });
       return createdDataset.id;
     },
-    delete: async (where) => {
+    delete: async (where, user) => {
       if (
         (where.id == null && where.slugs == null) ||
         (where.id != null && where.slugs != null)
@@ -133,12 +213,12 @@ export const repository: Repository = {
           "You should either specify the id or the slugs when deleting a dataset"
         );
       }
-
+      await checkUserAccessToDataset({ where, user });
       await prisma.dataset.delete({
         where: castObjectNullsToUndefined(where),
       });
     },
-    get: async (where) => {
+    get: async (where, user) => {
       if (
         (where.id == null && where.slugs == null) ||
         (where.id != null && where.slugs != null)
@@ -147,11 +227,13 @@ export const repository: Repository = {
           "You should either specify the id or the slugs when looking for a dataset"
         );
       }
-      return await prisma.dataset.findUnique({
+      await checkUserAccessToDataset({ where, user });
+      const dataset = await prisma.dataset.findUnique({
         where: castObjectNullsToUndefined(where),
       });
+      return dataset;
     },
-    update: async (where, dataset) => {
+    update: async (where, dataset, user) => {
       if (
         (where.id == null && where.slugs == null) ||
         (where.id != null && where.slugs != null)
@@ -160,6 +242,7 @@ export const repository: Repository = {
           "You should either specify the id or the slugs when updating a dataset"
         );
       }
+      await checkUserAccessToDataset({ where, user });
       try {
         await prisma.dataset.update({
           where: castObjectNullsToUndefined(where),
@@ -170,14 +253,26 @@ export const repository: Repository = {
         return false;
       }
     },
-    list: (_where, skip = undefined, first = undefined) =>
+    list: (where, skip = undefined, first = undefined) =>
       prisma.dataset.findMany(
         castObjectNullsToUndefined({
           orderBy: { createdAt: Prisma.SortOrder.asc },
           skip,
           take: first,
+          where: {
+            workspace: {
+              slug: where?.workspaceSlug,
+              memberships: { some: { userId: where?.user?.id } },
+            },
+          },
         })
       ),
+  },
+  workspace: {
+    add: addWorkspace,
+    get: getWorkspace,
+    list: listWorkspace,
+    update: updateWorkspace,
   },
   upload: {
     delete: deleteFromStorage,
